Protected against missin log file in reset

+++ src/lsst/cmservice/handlers/script_handler.py
@@ -604,7 +604,7 @@ class ScriptHandler(BaseScriptHandler):
         update_fields: dict[str, Any] = {}
         if to_status.value <= StatusEnum.prepared.value:
             update_fields["stamp_url"] = None
-            if script.log_url:
+            if script.log_url and os.path.exists(script.log_url):
                 os.unlink(script.log_url)
             if script.stamp_url and os.path.exists(script.stamp_url):
                 os.unlink(script.stamp_url)
